resources/v0.96/requirements.txt
# Core runtime
python-docx==1.2.0
pandas==2.3.3
numpy==2.3.4
pymupdf==1.26.6
requests==2.32.5
openpyxl==3.1.5
python-dotenv==1.2.1
gdown==5.2.0
packaging==24.2
PyPDF2==3.0.1
pyarrow==22.0.0
tenacity==9.1.2
lxml==6.0.2

# Gemini AI
google-generativeai==0.8.5

# Web interface
flask==3.1.2
flask-cors==6.0.1
streamlit==1.42.0
watchdog==6.0.0


# Progress tracking
tqdm==4.67.1
progressbar2==4.5.0

# Legacy .doc parsing is optional. Install textract==1.6.3 manually if needed.

# Testing
pytest==8.3.4
pytest-cov==6.0.0
